feat: created `GitHubRepositoryDTO`
